pyproject.toml
[tool.poetry]
name = "digimon"
version = "0.1.0"
description = ""
authors = ["Natthawut Suksai <owen55660@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
sqlmodel = "^0.0.21"
asyncpg = "^0.29.0"
pydantic-settings = "^2.4.0"
pyjwt = "^2.9.0"
pytest-mock = "^3.14.0"
bcrypt = "^4.2.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
passlib = "^1.7.4"


[tool.poetry.group.develop.dependencies]
pytest = "^8.3.2"
black = "^24.8.0"
httpx = "^0.27.0"
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.23.8"
locust = "^2.31.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
